Change Lighthouse config from CI file to config file.
Also changes Lighthouse CI to not check Cache TTL and Content Security Policy regarding XSS, as those depend on the server, and we run against a local NodeJS.
So any results will not be representative of production anyway.

The same goes for unminified-javascript as minification is done on the edge (by CloudFlare).

For now min-font-size is also disables, as Lighthous mistakes readable SVG text for unreadable.
(See https://github.com/GoogleChrome/lighthouse/issues/11286).